task6

+++ 0x05-pointers_arrays_strings/6-puts2.c
@@ -8,7 +8,7 @@
  */
 void puts2(char *str)
 {
-	int i;
+	long int i;
 
 	for (i = 0; i < strlen(str); i++)
 	{
